+++ src/types/product.ts
@@ -9,3 +9,16 @@ export interface ProductDetails {
   brand: string;
   rating: number;
 }
+
+export interface ImageDataType {
+  id: string;
+  index: number;
+  src: string;
+  productName: string;
+  actualPrice: number;
+  discountPrice: number;
+  discountPercentage: number;
+  size: string;
+  productDetails: string;
+  rating: number;
+}
